requirements.txt
mysqlclient
Django
djangorestframework
django-mysql
